requirements.txt
ahk==0.13.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
